Create ExternalStudio to allow for use in production environment.
The standard Studio allows for code injection as a feature. This makes it untenable to use on
anything other than your own machine.

This is a first stab at a public facing version. It replaces the result processor with a
Pandas specific query string. This should not allow for code injection.

More work needs to be done before I would be happy with this.