Add syntax highlighting and a proper code editor.
The result processor now uses a small ace-editor. It syntax highlights as well as debouncing so input is not sluggish.

The dash-ace dependency is currently pulled off of a fork in order to activate persistence, this shoud  move of fthe primary package once https://github.com/reasoned-ai/dash-ace/pull/2 is merged.

The code definition is now highlighted using pygments.